requirements.txt
Flask==3.1.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
flasgger==0.9.7.1
marshmallow==3.21.1
marshmallow-sqlalchemy==1.0.0
psycopg2-binary==2.9.9
python-dotenv==1.2.2
SQLAlchemy==2.0.30
